imp: make training code to be ddp-compatible
code in previous commit creates evaluation loss placeholder in the cpu, pull out
the loss value from singleton tensor in gpu and allocates the value in the placeholder.
this invokes error roots from device placement mismatch, whose error message is
'tensors must be cuda and dense'. this is resolved by this commit